Years in IT Fields: inclusive calendar-month counting + real decimal
Two follow-up fixes per the user's feedback on the previous commit's
numbers:

1. Jan 1 - Mar 31 was coming out to 2 months, not 3 — the old
   monthsBetween used exact-elapsed-time math (borrowed from
   formatDuration.js's "time ago" semantics), which only counts a month
   as complete once the end day-of-month reaches/passes the start day.
   Replaced with monthsTouched: (endYear-startYear)*12 +
   (endMonth-startMonth) + 1, floored at 1 — calendar months an entry
   touches, inclusive of both ends, matching how people actually describe
   a job's length ("Jan-Mar" reads as 3 months, not 2). This also fixes a
   short multi-day stint coming out to 0 and effectively vanishing from
   the total, which read as data being silently discarded.

2. "jangan bulat" — stopped flooring the total to a whole number; now
   rounds to 1 decimal place (Math.round(months/12*10)/10) instead.

Since `stats.value` is `integer not null` in the DB, this needed a schema
migration (supabase/migrations/20260906134524_allow_decimal_stat_values.sql,
widening it to numeric(6,1)) — statSchema's `.int()` constraint dropped
to match. I could NOT push this migration myself: `supabase db push`
failed with a 403 (this session's linked CLI credentials don't have
migration-apply privileges on the project, though read-only `gen types`
still works). The migration file is committed as the source of truth,
but someone with full project access needs to apply it — either by
running `npx supabase db push` from a machine logged into an account with
owner/admin access, or by pasting the migration's ALTER TABLE statement
directly into the Supabase Studio SQL Editor. Until that runs, Postgres
will silently round any decimal `value` back to a whole number on write
(no error), so the decimal display won't actually appear yet even after
this deploys and a resync is triggered.

Also smoothed the home page's count-up animation (fixed step count +
fractional increments) so a decimal target animates properly instead of
counting 1, 2, 3 and only snapping down to e.g. 2.8 on the last frame.

Hand-verified the new month counting against the real experience data:
PT Selaras Citra Terabit 3mo, Universitas Kristen Satya Wacana 20mo,
Boemisora 1mo (was 0), PT Sumber Alfaria Trijaya Tbk 12mo — total 36mo =
3.0 years (lands on a whole number with this data, even though the
computation itself is no longer floored).

+++ src/lib/validation/schemas.ts
@@ -32,7 +32,10 @@ export const bulkSkillNamesSchema = z.preprocess((v) => {
 
 export const statSchema = z.object({
 	label: z.string().trim().min(1, 'Label wajib diisi.').max(80),
-	value: z.coerce.number().int().min(0, 'Value tidak boleh negatif.')
+	// Not .int() — "Years in IT Fields" is auto-computed to 1 decimal place
+	// (see lib/server/autoStats.ts), and a manual edit on any stat should
+	// accept the same range the auto-computed ones can land on.
+	value: z.coerce.number().min(0, 'Value tidak boleh negatif.')
 });
 
 export const ROLE_TYPE_OPTIONS = ['Full-time', 'Part-time', 'Internship', 'Freelance', 'Project-based', 'Volunteer'];
